fix bug

+++ bin/tomo_viewer.py
@@ -141,7 +141,8 @@ class TomoViewer:
                 self.viewer.layers.move(self.viewer.layers.index(self.viewer.layers['ori_tomo']), 0)
                 self.viewer.add_points(name='edit vesicles', ndim=3, size=4)
                 self.viewer.layers['edit vesicles'].mode = 'ADD'
-                
+                self.multiple_viewer_widget.viewer_model1.camera.zoom = 0.9
+                self.multiple_viewer_widget.viewer_model2.camera.zoom = 0.9
                 self.progress_dialog.setValue(100)
                 message = f"Successfully opened the original image {file_path}."
                 self.print(message)
@@ -197,32 +198,36 @@ class TomoViewer:
         self.toolbar_widget.tabs.setCurrentIndex(1)
         
     def predict_clicked(self):
-        from qtpy.QtWidgets import QProgressDialog
-        from qtpy.QtCore import Qt
-        self.progress_dialog = QProgressDialog("Processing...", 'Cancel', 0, 100, self.main_viewer)
-        self.progress_dialog.setWindowTitle('Predicting')
-        self.progress_dialog.setWindowModality(Qt.WindowModal)
-        self.progress_dialog.setValue(0)
-        self.progress_dialog.show()
-        self.deconv_data = self.viewer.layers['deconv_tomo'].data
-        self.corrected_data = self.viewer.layers['corrected_tomo'].data
-        self.progress_dialog.setValue(20)
-        self.label = predict_label(self.deconv_data, self.corrected_data)
-        # self.label = self.viewer.layers['label'].data
-        self.progress_dialog.setValue(40)
-        self.area_path = self.tomo_path_and_stage.area_path
-        self.processed_vesicles, self.shape = morph_process(self.label, self.area_path)
-        self.progress_dialog.setValue(60)
+        if os.path.exists(self.tomo_path_and_stage.isonet_tomo_path):
+            from qtpy.QtWidgets import QProgressDialog
+            from qtpy.QtCore import Qt
+            self.progress_dialog = QProgressDialog("Processing...", 'Cancel', 0, 100, self.main_viewer)
+            self.progress_dialog.setWindowTitle('Predicting')
+            self.progress_dialog.setWindowModality(Qt.WindowModal)
+            self.progress_dialog.setValue(0)
+            self.progress_dialog.show()
+            self.deconv_data = self.viewer.layers['deconv_tomo'].data
+            self.corrected_data = self.viewer.layers['corrected_tomo'].data
+            self.progress_dialog.setValue(20)
+            self.label = predict_label(self.deconv_data, self.corrected_data)
+            # self.label = self.viewer.layers['label'].data
+            self.progress_dialog.setValue(40)
+            self.area_path = self.tomo_path_and_stage.area_path
+            self.processed_vesicles, self.shape = morph_process(self.label, self.area_path)
+            self.progress_dialog.setValue(60)
 
-        self.vesicle_info = vesicle_measure(self.corrected_data, self.processed_vesicles, self.shape, min_radius=8)
-        with open(self.tomo_path_and_stage.new_json_file_path,"w") as out:
-            json.dump(self.vesicle_info,out)
-        self.ves_tomo = vesicle_rendering(self.vesicle_info, self.shape)
-        with mrcfile.new(self.tomo_path_and_stage.new_label_file_path, overwrite=True) as mrc:
-            mrc.set_data(self.ves_tomo)
-        self.viewer.add_labels(self.ves_tomo, name='label')
-        self.viewer.layers['label'].opacity = 0.5 
-        self.progress_dialog.setValue(100)
+            self.vesicle_info = vesicle_measure(self.corrected_data, self.processed_vesicles, self.shape, min_radius=8)
+            with open(self.tomo_path_and_stage.new_json_file_path,"w") as out:
+                json.dump(self.vesicle_info,out)
+            self.ves_tomo = vesicle_rendering(self.vesicle_info, self.shape)
+            with mrcfile.new(self.tomo_path_and_stage.new_label_file_path, overwrite=True) as mrc:
+                mrc.set_data(self.ves_tomo)
+            self.viewer.add_labels(self.ves_tomo, name='label')
+            self.viewer.layers['label'].opacity = 0.5 
+            self.viewer.layers.selection.active = self.viewer.layers['edit vesicles']
+            self.progress_dialog.setValue(100)
+        else:
+            self.print("Please draw tomo area first.")
         
     def register_draw_area_mod(self):
